files: separate discretization from refinement tests
